# packages/gr-api/diagnosis-design.openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "持仓诊断 API 设计规范",
    "version": "0.3.0",
    "description": "根据《持仓诊断_表与接口设计.md》v0.3.0 生成的设计契约，尚不等同于已部署接口。保留设计路径 /api/v1/diagnosis 及直接响应体；当前 gr-api 实际路径为 /v1/diagnosis，且使用 code/message/data/timestamp/request_id 信封，仍需独立对齐实现。不得直接将此文件作为当前生产服务的已验证客户端契约。P2 分享端点保留已知定义，未决响应不臆造。"
  },
...
  "paths": {
    "/api/v1/diagnosis/snapshots": {
      "post": {
        "tags": [
          "持仓诊断"
        ],
        "operationId": "createDiagnosisSnapshot",
        "summary": "提交持仓诊断",
        "description": "新提交创建新 snapshot，计算结果仍可共享。只有相同 Idempotency-Key 重放原请求资源。所有成功状态使用 SnapshotAccepted，不直接返回完整计算结果。未完成为 202，新资源终态为 201，幂等重放终态为 200。",
        "responses": {
          "200": {
            "description": "幂等重放：原快照已到终态，包括 failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotAccepted"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Location": {
                "$ref": "#/components/headers/Location"
              }
            }
          },
          "201": {
            "description": "新建快照已到终态，包括 failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotAccepted"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Location": {
                "$ref": "#/components/headers/Location"
              }
            }
          },
          "202": {
            "description": "已接受，仍在 pending/running；新提交和重放均可返回。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotAccepted"
                },
                "example": {
                  "snapshot_id": "be6ebd1d-ab25-4ce5-8b8a-e366ab509f2f",
                  "schema_version": "diagnosis.v1",
                  "spec_version": "v1-hist-20260901",
                  "as_of_date": "2026-09-04",
                  "data_fingerprint": "20260904T1800",
                  "data_snapshot_at": "2026-09-04T18:00:00+08:00",
                  "resolved_intent": "single_portfolio",
                  "report_mode": "standard",
                  "requested_cov_methods": [
                    "historical"
                  ],
                  "run_status": "pending",
                  "plans": [
                    {
                      "plan_id": "p1",
                      "label": null,
                      "run_status": "pending",
                      "sections": {
                        "section_a": "pending",
                        "section_b": "pending",
                        "section_c": "pending",
                        "section_d": "pending",
                        "stress_scenarios": "pending",
                        "blindspots": "pending",
                        "profile": "pending"
                      },
                      "error": null
                    }
                  ],
                  "links": {
                    "result": "/api/v1/diagnosis/snapshots/be6ebd1d-ab25-4ce5-8b8a-e366ab509f2f/result",
                    "report": "/api/v1/diagnosis/snapshots/be6ebd1d-ab25-4ce5-8b8a-e366ab509f2f/report",
                    "stream": "/api/v1/diagnosis/snapshots/be6ebd1d-ab25-4ce5-8b8a-e366ab509f2f/stream",
                    "data_quality": "/api/v1/diagnosis/snapshots/be6ebd1d-ab25-4ce5-8b8a-e366ab509f2f/data-quality"
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Location": {
                "$ref": "#/components/headers/Location"
              },
              "Retry-After": {
                "$ref": "#/components/headers/RetryAfter"
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/IdempotencyConflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-stage": "P0",
        "x-design-source": "持仓诊断_表与接口设计.md §7",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "客户端安全随机生成的 UUIDv4。同一次操作的网络重试复用该键；重新诊断或重试计算使用新键。相同身份空间下同键异体返回 409。至少保留 24 小时，完整规则见 §7.8。",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotRequest"
              },
              "examples": {
                "weighted": {
                  "summary": "单方案显式权重",
                  "value": {
                    "plans": [
                      {
                        "plan_id": "p1",
                        "holdings": [
                          {
                            "symbol": "600000.SH",
                            "weight": "40"
                          },
                          {
                            "symbol": "000001.SZ",
                            "weight": "60"
                          }
                        ]
                      }
                    ],
                    "intent": "single_portfolio"
                  }
                },
                "equal": {
                  "summary": "先去重后等权",
                  "value": {
                    "plans": [
                      {
                        "plan_id": "p1",
                        "holdings": [
                          {
                            "symbol": "600000.SH"
                          },
                          {
                            "symbol": "000001.SZ"
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "x-access-control": "可匿名提交；已登录请求按用户隔离幂等空间，匿名请求使用独立匿名空间。"
      }
    },
    "/api/v1/diagnosis/snapshots/{id}/result": {
      "get": {
        "tags": [
          "持仓诊断"
        ],
        "operationId": "getDiagnosisResult",
        "summary": "读取全量诊断结果",
        "description": "从固定的 snapshot 终态读取；处理中返回进度，终态返回本端点模型。报告和数据质量均由同一份固定结果裁剪。",
        "responses": {
          "200": {
            "description": "已读取终态。HTTP 200 不代表诊断成功，必须检查 run_status；区块按 ready/unavailable/failed 分派。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResult"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "202": {
            "description": "计算尚未完成；按 Retry-After 轮询。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotProgress"
                },
                "example": {
                  "snapshot_id": "be6ebd1d-ab25-4ce5-8b8a-e366ab509f2f",
                  "schema_version": "diagnosis.v1",
                  "spec_version": "v1-hist-20260901",
                  "as_of_date": "2026-09-04",
                  "data_fingerprint": "20260904T1800",
                  "data_snapshot_at": "2026-09-04T18:00:00+08:00",
                  "resolved_intent": "single_portfolio",
                  "report_mode": "standard",
                  "requested_cov_methods": [
                    "historical"
                  ],
                  "run_status": "pending",
                  "plans": [
                    {
                      "plan_id": "p1",
                      "label": null,
                      "run_status": "pending",
                      "sections": {
                        "section_a": "pending",
                        "section_b": "pending",
                        "section_c": "pending",
                        "section_d": "pending",
                        "stress_scenarios": "pending",
                        "blindspots": "pending",
                        "profile": "pending"
                      },
                      "error": null
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Retry-After": {
                "$ref": "#/components/headers/RetryAfter"
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-stage": "P0",
        "x-design-source": "持仓诊断_表与接口设计.md §7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "UUIDv4 快照标识；匿名快照中同时是不透明访问凭证。",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            }
          }
        ],
        "x-access-control": "已登录快照校验请求者归属；匿名快照按 UUIDv4 snapshot_id 访问。鉴权传输方式尚未在设计稿中定义，参见 x-contract-gaps。"
      }
    },
    "/api/v1/diagnosis/snapshots/{id}/report": {
      "get": {
        "tags": [
          "持仓诊断"
        ],
        "operationId": "getDiagnosisReport",
        "summary": "读取按偏好裁剪的报告",
        "description": "从固定的 snapshot 终态读取；处理中返回进度，终态返回本端点模型。报告和数据质量均由同一份固定结果裁剪。",
        "responses": {
          "200": {
            "description": "已读取终态。HTTP 200 不代表诊断成功，必须检查 run_status；区块按 ready/unavailable/failed 分派。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                },
                "examples": {
                  "unavailableSectionShape": {
                    "summary": "完整最小结构示例；实际报告必须包含所有适用章节和强制声明",
                    "value": {
                      "snapshot_id": "be6ebd1d-ab25-4ce5-8b8a-e366ab509f2f",
                      "schema_version": "diagnosis.v1",
                      "spec_version": "v1-hist-20260901",
                      "as_of_date": "2026-09-04",
                      "data_fingerprint": "20260904T1800",
                      "data_snapshot_at": "2026-09-04T18:00:00+08:00",
                      "resolved_intent": "single_portfolio",
                      "report_mode": "standard",
                      "requested_cov_methods": [
                        "historical"
                      ],
                      "user_level": "retail",
                      "run_status": "partially_succeeded",
                      "sections": [
                        {
                          "section_id": "p1:style",
                          "plan_id": "p1",
                          "kind": "style",
                          "title": "风格与暴露",
                          "content": {
                            "status": "unavailable",
                            "reason_code": "factor_model_unavailable",
                            "reason": "因子数据尚不可用，暂时无法判断风格暴露。"
                          }
                        }
                      ],
                      "disclosures": [
                        "本报告基于 2026-09-04 18:00 的数据快照计算，数据供应商的后续修订不会反映在本报告中。"
                      ],
                      "data_quality": {
                        "plans": [
                          {
                            "plan_id": "p1",
                            "issues": [],
                            "coverage_summary": {
                              "input": 1.0
                            },
                            "forced_disclosures": []
                          }
                        ],
                        "forced_disclosures": [
                          "本报告不代表用户真实收益。"
                        ]
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateNoStore"
              }
            }
          },
          "202": {
            "description": "计算尚未完成；按 Retry-After 轮询。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotProgress"
                },
                "example": {
                  "snapshot_id": "be6ebd1d-ab25-4ce5-8b8a-e366ab509f2f",
                  "schema_version": "diagnosis.v1",
                  "spec_version": "v1-hist-20260901",
                  "as_of_date": "2026-09-04",
                  "data_fingerprint": "20260904T1800",
                  "data_snapshot_at": "2026-09-04T18:00:00+08:00",
                  "resolved_intent": "single_portfolio",
                  "report_mode": "standard",
                  "requested_cov_methods": [
                    "historical"
                  ],
                  "run_status": "pending",
                  "plans": [
                    {
                      "plan_id": "p1",
                      "label": null,
                      "run_status": "pending",
                      "sections": {
                        "section_a": "pending",
                        "section_b": "pending",
                        "section_c": "pending",
                        "section_d": "pending",
                        "stress_scenarios": "pending",
                        "blindspots": "pending",
                        "profile": "pending"
                      },
                      "error": null
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Retry-After": {
                "$ref": "#/components/headers/RetryAfter"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateNoStore"
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-stage": "P0",
        "x-design-source": "持仓诊断_表与接口设计.md §7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "UUIDv4 快照标识；匿名快照中同时是不透明访问凭证。",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            }
          },
          {
            "name": "user_level",
            "in": "query",
            "required": false,
            "description": "访问时的报告渲染偏好，与付费无关；切换不触发计算。",
            "schema": {
              "type": "string",
              "enum": [
                "retail",
                "pro"
              ],
              "default": "retail"
            }
          }
        ],
        "x-access-control": "已登录快照校验请求者归属；匿名快照按 UUIDv4 snapshot_id 访问。鉴权传输方式尚未在设计稿中定义，参见 x-contract-gaps。"
      }
    },
    "/api/v1/diagnosis/snapshots/{id}/data-quality": {
      "get": {
        "tags": [
          "持仓诊断"
        ],
        "operationId": "getDiagnosisDataQuality",
        "summary": "读取数据质量与强制声明",
        "description": "从固定的 snapshot 终态读取；处理中返回进度，终态返回本端点模型。报告和数据质量均由同一份固定结果裁剪。",
        "responses": {
          "200": {
            "description": "已读取终态。HTTP 200 不代表诊断成功，必须检查 run_status；区块按 ready/unavailable/failed 分派。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataQualityResponse"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "202": {
            "description": "计算尚未完成；按 Retry-After 轮询。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotProgress"
                },
                "example": {
                  "snapshot_id": "be6ebd1d-ab25-4ce5-8b8a-e366ab509f2f",
                  "schema_version": "diagnosis.v1",
                  "spec_version": "v1-hist-20260901",
                  "as_of_date": "2026-09-04",
                  "data_fingerprint": "20260904T1800",
                  "data_snapshot_at": "2026-09-04T18:00:00+08:00",
                  "resolved_intent": "single_portfolio",
                  "report_mode": "standard",
                  "requested_cov_methods": [
                    "historical"
                  ],
                  "run_status": "pending",
                  "plans": [
                    {
                      "plan_id": "p1",
                      "label": null,
                      "run_status": "pending",
                      "sections": {
                        "section_a": "pending",
                        "section_b": "pending",
                        "section_c": "pending",
                        "section_d": "pending",
                        "stress_scenarios": "pending",
                        "blindspots": "pending",
                        "profile": "pending"
                      },
                      "error": null
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Retry-After": {
                "$ref": "#/components/headers/RetryAfter"
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-stage": "P0",
        "x-design-source": "持仓诊断_表与接口设计.md §7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "UUIDv4 快照标识；匿名快照中同时是不透明访问凭证。",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            }
          }
        ],
        "x-access-control": "已登录快照校验请求者归属；匿名快照按 UUIDv4 snapshot_id 访问。鉴权传输方式尚未在设计稿中定义，参见 x-contract-gaps。"
      }
    },
    "/api/v1/diagnosis/spec-versions": {
      "get": {
        "tags": [
          "持仓诊断"
        ],
        "operationId": "listDiagnosisSpecVersions",
        "summary": "列出口径版本",
        "description": "按 effective_from 降序一次返回全部版本。",
        "responses": {
          "200": {
            "description": "版本列表。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecVersionList"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-stage": "P0",
        "x-design-source": "持仓诊断_表与接口设计.md §7"
      }
    },
    "/api/v1/diagnosis/spec-versions/current": {
      "get": {
        "tags": [
          "持仓诊断"
        ],
        "operationId": "getCurrentDiagnosisSpecVersion",
        "summary": "读取当前口径配置",
        "description": "没有当前生效版本时返回 503。配置值不等于已完成计算的方法。",
        "responses": {
          "200": {
            "description": "当前生效版本及配置。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecVersionInfo"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-stage": "P0",
        "x-design-source": "持仓诊断_表与接口设计.md §7"
      }
    },
...
  },
  "components": {
...
    "headers": {
      "RequestId": {
        "description": "与 JSON ErrorResponse.request_id 相同的请求追踪标识。",
        "schema": {
          "type": "string"
        }
      },
      "RetryAfter": {
        "description": "建议等待秒数；设计一期为 1。",
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "example": 1
      },
      "Location": {
        "description": "指向响应 links.result 的同源相对地址。",
        "schema": {
          "type": "string",
          "format": "uri-reference"
        }
      },
      "PrivateNoStore": {
        "description": "报告按访问偏好生成，不存储 HTTP 缓存。",
        "schema": {
          "type": "string",
          "const": "private, no-store"
        }
      }
    },
    "responses": {
      "InvalidRequest": {
        "description": "请求、路径或查询参数校验失败；issues.path 使用原始输入位置。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
      "IdempotencyConflict": {
        "description": "相同身份空间中的幂等键已对应不同请求体。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "request_id": "example-request-id",
              "error": {
                "code": "IDEMPOTENCY_KEY_CONFLICT",
                "message": "同一幂等键已用于不同请求。",
                "retryable": false,
                "issues": []
              }
            }
          }
        },
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
...
    }
  },
...
}